+++ shacl_ast/src/ast/node_shape.rs
@@ -1,8 +1,9 @@
 use crate::shacl_vocab::{
-    sh_closed, sh_deactivated, sh_description, sh_group, sh_info, sh_name, sh_node_shape,
-    sh_property, sh_severity, sh_violation, sh_warning,
+    sh_closed, sh_description, sh_group, sh_info, sh_name, sh_node_shape, sh_property, sh_severity,
+    sh_violation, sh_warning,
 };
 use crate::{component::Component, message_map::MessageMap, severity::Severity, target::Target};
+use prefixmap::IriRef;
 use srdf::{BuildRDF, RDFNode, Rdf};
 use std::fmt::Display;
 
@@ -15,9 +16,8 @@ where
     components: Vec<Component>,
     targets: Vec<Target<RDF>>,
     property_shapes: Vec<RDFNode>,
-    closed: bool,
+    // closed: bool,
     // ignored_properties: Vec<IriRef>,
-    deactivated: bool,
     // message: MessageMap,
     severity: Option<Severity>,
     name: MessageMap,
@@ -33,9 +33,8 @@ impl<RDF: Rdf> NodeShape<RDF> {
             components: Vec::new(),
             targets: Vec::new(),
             property_shapes: Vec::new(),
-            closed: false,
+            // closed: false,
             // ignored_properties: Vec::new(),
-            deactivated: false,
             // message: MessageMap::new(),
             severity: None,
             name: MessageMap::new(),
@@ -64,21 +63,30 @@ impl<RDF: Rdf> NodeShape<RDF> {
         self
     }
 
-    pub fn with_closed(mut self, closed: bool) -> Self {
-        self.closed = closed;
-        self
-    }
-
     pub fn id(&self) -> &RDFNode {
         &self.id
     }
 
-    pub fn is_closed(&self) -> &bool {
-        &self.closed
+    pub fn is_deactivated(&self) -> bool {
+        for component in &self.components {
+            if let Component::Deactivated(true) = component {
+                return true;
+            }
+        }
+        return false;
     }
 
-    pub fn is_deactivated(&self) -> &bool {
-        &self.deactivated
+    pub fn closed_component(&self) -> (bool, Vec<IriRef>) {
+        for component in &self.components {
+            if let Component::Closed {
+                is_closed,
+                ignored_properties,
+            } = component
+            {
+                return (*is_closed, ignored_properties.clone());
+            }
+        }
+        return (false, Vec::new());
     }
 
     pub fn severity(&self) -> Option<Severity> {
@@ -133,12 +141,6 @@ impl<RDF: Rdf> NodeShape<RDF> {
             rdf.add_triple(id.clone(), sh_property().clone(), property_shape.clone())
         })?;
 
-        if self.deactivated {
-            let literal: B::Literal = "true".to_string().into();
-
-            rdf.add_triple(id.clone(), sh_deactivated().clone(), literal)?;
-        }
-
         if let Some(group) = &self.group {
             rdf.add_triple(id.clone(), sh_group().clone(), group.clone())?;
         }
@@ -154,12 +156,6 @@ impl<RDF: Rdf> NodeShape<RDF> {
             rdf.add_triple(id.clone(), sh_severity().clone(), pred.clone())?;
         }
 
-        if self.closed {
-            let literal: B::Literal = "true".to_string().into();
-
-            rdf.add_triple(id.clone(), sh_closed().clone(), literal)?;
-        }
-
         Ok(())
     }
 }
@@ -167,9 +163,6 @@ impl<RDF: Rdf> NodeShape<RDF> {
 impl<RDF: Rdf> Display for NodeShape<RDF> {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         writeln!(f, "{{")?;
-        if self.closed {
-            writeln!(f, "       closed: {}", self.closed)?
-        }
         for target in self.targets.iter() {
             writeln!(f, "       {target}")?
         }
@@ -191,8 +184,6 @@ impl<RDF: Rdf> Clone for NodeShape<RDF> {
             components: self.components.clone(),
             targets: self.targets.clone(),
             property_shapes: self.property_shapes.clone(),
-            closed: self.closed,
-            deactivated: self.deactivated,
             severity: self.severity.clone(),
             name: self.name.clone(),
             description: self.description.clone(),
@@ -207,8 +198,6 @@ impl<RDF: Rdf> PartialEq for NodeShape<RDF> {
             && self.components == other.components
             && self.targets == other.targets
             && self.property_shapes == other.property_shapes
-            && self.closed == other.closed
-            && self.deactivated == other.deactivated
             && self.severity == other.severity
             && self.name == other.name
             && self.description == other.description