+++ shacl_ir/src/compiled/node_shape.rs
@@ -5,6 +5,8 @@ use srdf::{RDFNode, Rdf};
 use shacl_ast::node_shape::NodeShape;
 use shacl_ast::Schema;
 
+use crate::closed_info::ClosedInfo;
+
 use super::compile_shape;
 use super::compiled_shacl_error::CompiledShaclError;
 use super::component::CompiledComponent;
@@ -18,8 +20,7 @@ pub struct CompiledNodeShape {
     components: Vec<CompiledComponent>,
     targets: Vec<CompiledTarget>,
     property_shapes: Vec<CompiledShape>,
-    closed: bool,
-    // ignored_properties: Vec<S::IRI>,
+    closed_info: ClosedInfo,
     deactivated: bool,
     // message: MessageMap,
     severity: Option<CompiledSeverity>,
@@ -35,7 +36,7 @@ impl CompiledNodeShape {
         components: Vec<CompiledComponent>,
         targets: Vec<CompiledTarget>,
         property_shapes: Vec<CompiledShape>,
-        closed: bool,
+        closed_info: ClosedInfo,
         deactivated: bool,
         severity: Option<CompiledSeverity>,
     ) -> Self {
@@ -44,7 +45,7 @@ impl CompiledNodeShape {
             components,
             targets,
             property_shapes,
-            closed,
+            closed_info: ClosedInfo::default(),
             deactivated,
             severity,
         }
@@ -77,26 +78,28 @@ impl CompiledNodeShape {
         &self.property_shapes
     }
 
-    pub fn closed(&self) -> &bool {
-        &self.closed
+    pub fn closed(&self) -> bool {
+        self.closed_info.is_closed()
     }
 }
 
 impl CompiledNodeShape {
+    /// Compiles an AST NodeShape to an internal representation NodeShape
+    /// It embeds some components like deactivated as boolean attributes of the internal representation of the node shape
     pub fn compile<S: Rdf>(
         shape: Box<NodeShape<S>>,
         schema: &Schema<S>,
     ) -> Result<Self, CompiledShaclError> {
         let id = shape.id().clone();
-        let closed = shape.is_closed().to_owned();
         let deactivated = shape.is_deactivated().to_owned();
         let severity = CompiledSeverity::compile(shape.severity())?;
 
         let components = shape.components().iter().collect::<HashSet<_>>();
         let mut compiled_components = Vec::new();
         for component in components {
-            let component = CompiledComponent::compile(component.to_owned(), schema)?;
-            compiled_components.push(component);
+            if let Some(component) = CompiledComponent::compile(component.to_owned(), schema)? {
+                compiled_components.push(component);
+            }
         }
 
         let mut targets = Vec::new();
@@ -111,12 +114,14 @@ impl CompiledNodeShape {
             property_shapes.push(shape);
         }
 
+        let closed_info = ClosedInfo::get_closed_info_node_shape(&shape);
+
         let compiled_node_shape = CompiledNodeShape::new(
             id,
             compiled_components,
             targets,
             property_shapes,
-            closed,
+            closed_info,
             deactivated,
             severity,
         );