+++ shex_ast/src/ir/value_set_value.rs
@@ -88,6 +88,7 @@ impl ValueSetValue {
                         None => false,
                     },
                     srdf::SLiteral::DatatypeLiteral { .. } => false,
+                    srdf::SLiteral::WrongDatatypeLiteral { .. } => false,
                     srdf::SLiteral::NumericLiteral(_) => false,
                     srdf::SLiteral::DatetimeLiteral(_) => false,
                     srdf::SLiteral::BooleanLiteral(_) => false,